code/backend/requirements.txt
asgiref
celery
certifi
charset-normalizer
contextlib2
cryptography
Django
django-cors-headers
django-redis
drf-spectacular
et-xmlfile
idna
mkdocs
mkdocs-material
mysqlclient
numpy
openpyxl
pandas
pika
pycryptodome
pydantic
PyJWT
pymdown-extensions
pymysql
pypinyin
python-dateutil
python-ldap
pytz
PyYAML
requests
schema
six
sqlparse
typing_extensions
tzdata
urllib3
uuid
uwsgi
xmltodict
fastapi
uvicorn[standard]
python-dotenv
httpx
pytest
websockets
gunicorn